task should be created with open state
