+++ todo/stories_test.py
@@ -30,7 +30,7 @@ async def test_new_task_story(build_context, mocker):
 
         assert obj['list'] == 'list_1'
         assert obj['description'] == 'hello, world!'
-        assert obj['state'] == 'new'
+        assert obj['state'] == 'open'
         assert 'created_at' in obj
         assert 'updated_at' in obj
 
@@ -573,35 +573,3 @@ async def test_remove_task_by_postback_fail_if_wrong_id(build_context):
         ])
         tasks_left = await tasks_document.TaskDocument.objects.find()
         assert len(tasks_left) == 3
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(('task_idx', 'command_tmpl', 'should_get_answer', 'should_get_state'), [
-    (0, 'REOPEN_TASK_{}', ':ok: Task `{}` was opened', 'open'),
-    (0, 'DONE_TASK_{}', 'Task `{}` is already done', 'done'),
-    (1, 'STOP_TASK_{}', ':ok: Task `{}` was stopped', 'open'),
-    (1, 'DONE_TASK_{}', ':ok: Task `{}` was done', 'done'),
-    (1, 'START_TASK_{}', 'Task `{}` is already in progress', 'in progress'),
-    (2, 'START_TASK_{}', ':ok: Task `{}` was started', 'in progress'),
-    (2, 'REOPEN_TASK_{}', 'Task `{}` is already opened', 'open'),
-    (2, 'STOP_TASK_{}', 'Task `{}` is already opened', 'open'),
-])
-async def test_start_task_by_postback(
-        build_context, task_idx, command_tmpl, should_get_answer, should_get_state):
-    async with build_context() as ctx:
-        created_tasks = await ctx.add_test_tasks()
-
-        target_task_id = created_tasks[task_idx]._id
-
-        # Alice:
-        await ctx.dialog([
-            env.build_postback(command_tmpl.format(target_task_id)),
-        ])
-
-        task_after_command = await tasks_document.TaskDocument.objects.find_by_id(target_task_id)
-        # Bob:
-        await ctx.dialog([
-            None,
-            should_get_answer.format(task_after_command.description),
-        ])
-        assert task_after_command.state == should_get_state
